pyproject.toml
[project]
name = "spotify2yt"
version = "2.0"
description = "CLI tool to migrate playlists between Spotify and YouTube Music"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "spotipy",
    "ytmusicapi",
    "python-dotenv",
    "rich"
]

[project.scripts]
spotify2yt = "spotify2yt.cli:app"